pyproject.toml
[tool.poetry]
name = "pyopenaireq"
version = "0.1.0"
description = ""
authors = ["Shuichi Kataoka <shu01k9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
